+++ src/commonTest/kotlin/csense/kotlin/specificExtensions/boolean/BooleanMappingsTest.kt
@@ -1,10 +1,13 @@
 package csense.kotlin.specificExtensions.boolean
 
+import csense.kotlin.tests.assertions.*
 import kotlin.test.*
 
 class BooleanMappingsTest {
     @Test
-    fun booleanMappings(){
-        TODO()
+    fun booleanMappings() {
+        val first = true.mappings
+        val second = false.mappings
+        (first != second).assertTrue("should have different instances")
     }
 }